Add crash handler and error reporter to the project.
Using the Google-Cloud-CPP library to upload the error reports to google storage, the ProcessDump executable to dump the memory of the process during the crash, and the StackWalker library to get a stack trace of the crash.

Also added a basic winforms windows for the crash reporter, which requires grabbing a handle to the d3d11 module.

Also: add Git-LFS to handle the libs and exe's we're tracking.